+++ SpatialGDK/Source/SpatialGDK/Private/Utils/SpatialStatics.cpp
@@ -84,8 +84,9 @@ bool USpatialStatics::IsActorGroupOwnerForActor(const AActor* Actor)
 		return false;
 	}
 
+	// Offloading using the Unreal Load Balancing always load balances based on the owning actor.
 	const AActor* RootOwner = Actor;
-	while (RootOwner->bUseNetOwnerActorGroup && RootOwner->GetOwner() != nullptr)
+	while (RootOwner->GetOwner() != nullptr && RootOwner->GetOwner()->GetIsReplicated())
 	{
 		RootOwner = RootOwner->GetOwner();
 	}